requirements1.txt
random
cv2 (OpenCV)
numpy
ultralytics.YOLO
streamlit
PIL.Image
tempfile
base64